Project/cloud_run/new_script_service/requirements.txt
Flask==1.1.1
pytest==5.3.2; python_version > "3.0"
pytest==5.3.2; python_version < "3.0"
gunicorn==20.0.4
google-cloud-storage
pyinstaller
pipreqs